refactor: rename package "jsonutil" to "producer"
